Fix some minor issues with FF
