+++ packages/react/src/Breadcrumbs/Breadcrumbs.module.css
@@ -59,9 +59,21 @@
 .Item {
   display: inline-block;
   font-size: var(--text-body-size-medium);
+  color: var(--fgColor-link);
+  text-decoration: none;
 
-  &:focus-visible {
-    @mixin focusOutline;
+  &:hover,
+  &:focus {
+    text-decoration: underline;
+  }
+}
+
+.ItemSelected {
+  color: var(--fgColor-default);
+  pointer-events: none;
+
+  &:focus {
+    text-decoration: none;
   }
 }
 
@@ -77,6 +89,7 @@
     }
 
     &:focus-visible {
+      @mixin focusOutline 1px;
       text-decoration: none;
     }
 
@@ -97,6 +110,10 @@
     &:hover {
       background: var(--control-transparent-bgColor-hover);
       text-decoration: none;
+    }
+
+    &:focus-visible {
+      @mixin focusOutline;
     }
 
     &[aria-current] {
@@ -136,7 +153,6 @@
     & summary {
       list-style: none;
       cursor: pointer;
-      outline: none;
 
       &::-webkit-details-marker {
         display: none;
